Fixing an error where there were no validation if the data exists before trying to updated it.

+++ src/main/java/com/wuwa/helper/services/WeaponTypeService.java
@@ -43,7 +43,9 @@ public class WeaponTypeService {
     }
 
     public void deleteWeaponTypeById(String weaponTypeId){
-        weaponTypeRepository.deleteById(UUID.fromString(weaponTypeId));
+        if(weaponTypeExists(weaponTypeId)){
+            weaponTypeRepository.deleteById(UUID.fromString(weaponTypeId));
+        }
     }
 
     public WeaponType updateWeaponType(String weaponTypeId, WeaponTypeDTO weaponTypeDTO){
